// ui/src/types/sourcenetwork/acp_core/parser_message.ts
export interface LocatedMessage {
  message: string;
  kind: LocatedMessage_MessageKind;
  inputName: string;
  interval: BufferInterval | undefined;
}

export enum LocatedMessage_MessageKind {
  UNKNOWN = 0,
  ERROR = 1,
  INFO = 2,
  WARNING = 3,
  UNRECOGNIZED = -1,
}
// ...
export const LocatedMessage: MessageFns<LocatedMessage> = {
  fromJSON(object: any): LocatedMessage {
    return {
      message: isSet(object.message) ? globalThis.String(object.message) : "",
      kind: isSet(object.kind) ? locatedMessage_MessageKindFromJSON(object.kind) : 0,
      inputName: isSet(object.inputName) ? globalThis.String(object.inputName) : "",
      interval: isSet(object.interval) ? BufferInterval.fromJSON(object.interval) : undefined,
    };
  },

  toJSON(message: LocatedMessage): unknown {
    const obj: any = {};
    if (message.message !== "") {
      obj.message = message.message;
    }
    if (message.kind !== 0) {
      obj.kind = locatedMessage_MessageKindToJSON(message.kind);
    }
    if (message.inputName !== "") {
      obj.inputName = message.inputName;
    }
    if (message.interval !== undefined) {
      obj.interval = BufferInterval.toJSON(message.interval);
    }
    return obj;
  },

  create<I extends Exact<DeepPartial<LocatedMessage>, I>>(base?: I): LocatedMessage {
    return LocatedMessage.fromPartial(base ?? ({} as any));
  },
  fromPartial<I extends Exact<DeepPartial<LocatedMessage>, I>>(object: I): LocatedMessage {
    const message = createBaseLocatedMessage();
    message.message = object.message ?? "";
    message.kind = object.kind ?? 0;
    message.inputName = object.inputName ?? "";
    message.interval = (object.interval !== undefined && object.interval !== null)
      ? BufferInterval.fromPartial(object.interval)
      : undefined;
    return message;
  },
};

type Builtin = Date | Function | Uint8Array | string | number | boolean | undefined;

export type DeepPartial<T> = T extends Builtin ? T
  : T extends globalThis.Array<infer U> ? globalThis.Array<DeepPartial<U>>
  : T extends ReadonlyArray<infer U> ? ReadonlyArray<DeepPartial<U>>
  : T extends {} ? { [K in keyof T]?: DeepPartial<T[K]> }
  : Partial<T>;

type KeysOfUnion<T> = T extends T ? keyof T : never;
export type Exact<P, I extends P> = P extends Builtin ? P
  : P & { [K in keyof P]: Exact<P[K], I[K]> } & { [K in Exclude<keyof I, KeysOfUnion<P>>]: never };

function isSet(value: any): boolean {
  return value !== null && value !== undefined;
}

export interface MessageFns<T> {
  fromJSON(object: any): T;
  toJSON(message: T): unknown;
  create<I extends Exact<DeepPartial<T>, I>>(base?: I): T;
  fromPartial<I extends Exact<DeepPartial<T>, I>>(object: I): T;
}