// ui/src/types/sourcenetwork/acp_core/theorem.ts
export interface AuthorizationTheorem {
  operation: Operation | undefined;
  actor: Actor | undefined;
  assertTrue: boolean;
}

export interface ReachabilityTheorem {
  actor: Actor | undefined;
  operation: string;
  object: Object | undefined;
  assertTrue: boolean;
}

/**
 * DelegationTheorem models a statement declaring that actor can create and delete relationships
 * with `relation` for the given `object`
 */
export interface DelegationTheorem {
  actor: Actor | undefined;
  operation: Operation | undefined;
  assertTrue: boolean;
}

/** PolicyTheorem models a set of theorems about a Policy */
export interface PolicyTheorem {
  authorizationTheorems: AuthorizationTheorem[];
  delegationTheorems: DelegationTheorem[];
  reachabilityTheorems: ReachabilityTheorem[];
}

/** Result models the result of evaluating a theorem */
export interface Result {
  /** status indicates whether the theorem was accepted, rejected or an error happened */
  status: ResultStatus;
  /** Message explains the result */
  message: string;
}

export interface AuthorizationTheoremResult {
  theorem: AuthorizationTheorem | undefined;
  result: Result | undefined;
}

export interface DelegationTheoremResult {
  theorem: DelegationTheorem | undefined;
  result: Result | undefined;
}

export interface ReachabilityTheoremResult {
  theorem: ReachabilityTheorem | undefined;
  result: Result | undefined;
}

export interface AnnotatedAuthorizationTheoremResult {
  result: AuthorizationTheoremResult | undefined;
  interval: BufferInterval | undefined;
}

export interface AnnotatedDelegationTheoremResult {
  result: DelegationTheoremResult | undefined;
  interval: BufferInterval | undefined;
}
// ...
export const AuthorizationTheorem: MessageFns<AuthorizationTheorem> = {
  fromJSON(object: any): AuthorizationTheorem {
    return {
      operation: isSet(object.operation) ? Operation.fromJSON(object.operation) : undefined,
      actor: isSet(object.actor) ? Actor.fromJSON(object.actor) : undefined,
      assertTrue: isSet(object.assertTrue) ? globalThis.Boolean(object.assertTrue) : false,
    };
  },

  toJSON(message: AuthorizationTheorem): unknown {
    const obj: any = {};
    if (message.operation !== undefined) {
      obj.operation = Operation.toJSON(message.operation);
    }
    if (message.actor !== undefined) {
      obj.actor = Actor.toJSON(message.actor);
    }
    if (message.assertTrue !== false) {
      obj.assertTrue = message.assertTrue;
    }
    return obj;
  },

  create<I extends Exact<DeepPartial<AuthorizationTheorem>, I>>(base?: I): AuthorizationTheorem {
    return AuthorizationTheorem.fromPartial(base ?? ({} as any));
  },
  fromPartial<I extends Exact<DeepPartial<AuthorizationTheorem>, I>>(object: I): AuthorizationTheorem {
    const message = createBaseAuthorizationTheorem();
    message.operation = (object.operation !== undefined && object.operation !== null)
      ? Operation.fromPartial(object.operation)
      : undefined;
    message.actor = (object.actor !== undefined && object.actor !== null) ? Actor.fromPartial(object.actor) : undefined;
    message.assertTrue = object.assertTrue ?? false;
    return message;
  },
};
// ...
export const DelegationTheorem: MessageFns<DelegationTheorem> = {
  fromJSON(object: any): DelegationTheorem {
    return {
      actor: isSet(object.actor) ? Actor.fromJSON(object.actor) : undefined,
      operation: isSet(object.operation) ? Operation.fromJSON(object.operation) : undefined,
      assertTrue: isSet(object.assertTrue) ? globalThis.Boolean(object.assertTrue) : false,
    };
  },

  toJSON(message: DelegationTheorem): unknown {
    const obj: any = {};
    if (message.actor !== undefined) {
      obj.actor = Actor.toJSON(message.actor);
    }
    if (message.operation !== undefined) {
      obj.operation = Operation.toJSON(message.operation);
    }
    if (message.assertTrue !== false) {
      obj.assertTrue = message.assertTrue;
    }
    return obj;
  },

  create<I extends Exact<DeepPartial<DelegationTheorem>, I>>(base?: I): DelegationTheorem {
    return DelegationTheorem.fromPartial(base ?? ({} as any));
  },
  fromPartial<I extends Exact<DeepPartial<DelegationTheorem>, I>>(object: I): DelegationTheorem {
    const message = createBaseDelegationTheorem();
    message.actor = (object.actor !== undefined && object.actor !== null) ? Actor.fromPartial(object.actor) : undefined;
    message.operation = (object.operation !== undefined && object.operation !== null)
      ? Operation.fromPartial(object.operation)
      : undefined;
    message.assertTrue = object.assertTrue ?? false;
    return message;
  },
};
// ...
export const Result: MessageFns<Result> = {
  fromJSON(object: any): Result {
    return {
      status: isSet(object.status) ? resultStatusFromJSON(object.status) : 0,
      message: isSet(object.message) ? globalThis.String(object.message) : "",
    };
  },

  toJSON(message: Result): unknown {
    const obj: any = {};
    if (message.status !== 0) {
      obj.status = resultStatusToJSON(message.status);
    }
    if (message.message !== "") {
      obj.message = message.message;
    }
    return obj;
  },

  create<I extends Exact<DeepPartial<Result>, I>>(base?: I): Result {
    return Result.fromPartial(base ?? ({} as any));
  },
  fromPartial<I extends Exact<DeepPartial<Result>, I>>(object: I): Result {
    const message = createBaseResult();
    message.status = object.status ?? 0;
    message.message = object.message ?? "";
    return message;
  },
};

function createBaseAuthorizationTheoremResult(): AuthorizationTheoremResult {
  return { theorem: undefined, result: undefined };
}

export const AuthorizationTheoremResult: MessageFns<AuthorizationTheoremResult> = {
  fromJSON(object: any): AuthorizationTheoremResult {
    return {
      theorem: isSet(object.theorem) ? AuthorizationTheorem.fromJSON(object.theorem) : undefined,
      result: isSet(object.result) ? Result.fromJSON(object.result) : undefined,
    };
  },

  toJSON(message: AuthorizationTheoremResult): unknown {
    const obj: any = {};
    if (message.theorem !== undefined) {
      obj.theorem = AuthorizationTheorem.toJSON(message.theorem);
    }
    if (message.result !== undefined) {
      obj.result = Result.toJSON(message.result);
    }
    return obj;
  },

  create<I extends Exact<DeepPartial<AuthorizationTheoremResult>, I>>(base?: I): AuthorizationTheoremResult {
    return AuthorizationTheoremResult.fromPartial(base ?? ({} as any));
  },
  fromPartial<I extends Exact<DeepPartial<AuthorizationTheoremResult>, I>>(object: I): AuthorizationTheoremResult {
    const message = createBaseAuthorizationTheoremResult();
    message.theorem = (object.theorem !== undefined && object.theorem !== null)
      ? AuthorizationTheorem.fromPartial(object.theorem)
      : undefined;
    message.result = (object.result !== undefined && object.result !== null)
      ? Result.fromPartial(object.result)
      : undefined;
    return message;
  },
};

function createBaseDelegationTheoremResult(): DelegationTheoremResult {
  return { theorem: undefined, result: undefined };
}

export const DelegationTheoremResult: MessageFns<DelegationTheoremResult> = {
  fromJSON(object: any): DelegationTheoremResult {
    return {
      theorem: isSet(object.theorem) ? DelegationTheorem.fromJSON(object.theorem) : undefined,
      result: isSet(object.result) ? Result.fromJSON(object.result) : undefined,
    };
  },

  toJSON(message: DelegationTheoremResult): unknown {
    const obj: any = {};
    if (message.theorem !== undefined) {
      obj.theorem = DelegationTheorem.toJSON(message.theorem);
    }
    if (message.result !== undefined) {
      obj.result = Result.toJSON(message.result);
    }
    return obj;
  },

  create<I extends Exact<DeepPartial<DelegationTheoremResult>, I>>(base?: I): DelegationTheoremResult {
    return DelegationTheoremResult.fromPartial(base ?? ({} as any));
  },
  fromPartial<I extends Exact<DeepPartial<DelegationTheoremResult>, I>>(object: I): DelegationTheoremResult {
    const message = createBaseDelegationTheoremResult();
    message.theorem = (object.theorem !== undefined && object.theorem !== null)
      ? DelegationTheorem.fromPartial(object.theorem)
      : undefined;
    message.result = (object.result !== undefined && object.result !== null)
      ? Result.fromPartial(object.result)
      : undefined;
    return message;
  },
};
// ...
export const AnnotatedAuthorizationTheoremResult: MessageFns<AnnotatedAuthorizationTheoremResult> = {
  fromJSON(object: any): AnnotatedAuthorizationTheoremResult {
    return {
      result: isSet(object.result) ? AuthorizationTheoremResult.fromJSON(object.result) : undefined,
      interval: isSet(object.interval) ? BufferInterval.fromJSON(object.interval) : undefined,
    };
  },

  toJSON(message: AnnotatedAuthorizationTheoremResult): unknown {
    const obj: any = {};
    if (message.result !== undefined) {
      obj.result = AuthorizationTheoremResult.toJSON(message.result);
    }
    if (message.interval !== undefined) {
      obj.interval = BufferInterval.toJSON(message.interval);
    }
    return obj;
  },

  create<I extends Exact<DeepPartial<AnnotatedAuthorizationTheoremResult>, I>>(
    base?: I,
  ): AnnotatedAuthorizationTheoremResult {
    return AnnotatedAuthorizationTheoremResult.fromPartial(base ?? ({} as any));
  },
  fromPartial<I extends Exact<DeepPartial<AnnotatedAuthorizationTheoremResult>, I>>(
    object: I,
  ): AnnotatedAuthorizationTheoremResult {
    const message = createBaseAnnotatedAuthorizationTheoremResult();
    message.result = (object.result !== undefined && object.result !== null)
      ? AuthorizationTheoremResult.fromPartial(object.result)
      : undefined;
    message.interval = (object.interval !== undefined && object.interval !== null)
      ? BufferInterval.fromPartial(object.interval)
      : undefined;
    return message;
  },
};

function createBaseAnnotatedDelegationTheoremResult(): AnnotatedDelegationTheoremResult {
  return { result: undefined, interval: undefined };
}

export const AnnotatedDelegationTheoremResult: MessageFns<AnnotatedDelegationTheoremResult> = {
  fromJSON(object: any): AnnotatedDelegationTheoremResult {
    return {
      result: isSet(object.result) ? DelegationTheoremResult.fromJSON(object.result) : undefined,
      interval: isSet(object.interval) ? BufferInterval.fromJSON(object.interval) : undefined,
    };
  },

  toJSON(message: AnnotatedDelegationTheoremResult): unknown {
    const obj: any = {};
    if (message.result !== undefined) {
      obj.result = DelegationTheoremResult.toJSON(message.result);
    }
    if (message.interval !== undefined) {
      obj.interval = BufferInterval.toJSON(message.interval);
    }
    return obj;
  },

  create<I extends Exact<DeepPartial<AnnotatedDelegationTheoremResult>, I>>(
    base?: I,
  ): AnnotatedDelegationTheoremResult {
    return AnnotatedDelegationTheoremResult.fromPartial(base ?? ({} as any));
  },
  fromPartial<I extends Exact<DeepPartial<AnnotatedDelegationTheoremResult>, I>>(
    object: I,
  ): AnnotatedDelegationTheoremResult {
    const message = createBaseAnnotatedDelegationTheoremResult();
    message.result = (object.result !== undefined && object.result !== null)
      ? DelegationTheoremResult.fromPartial(object.result)
      : undefined;
    message.interval = (object.interval !== undefined && object.interval !== null)
      ? BufferInterval.fromPartial(object.interval)
      : undefined;
    return message;
  },
};
// ...
type Builtin = Date | Function | Uint8Array | string | number | boolean | undefined;

export type DeepPartial<T> = T extends Builtin ? T
  : T extends globalThis.Array<infer U> ? globalThis.Array<DeepPartial<U>>
  : T extends ReadonlyArray<infer U> ? ReadonlyArray<DeepPartial<U>>
  : T extends {} ? { [K in keyof T]?: DeepPartial<T[K]> }
  : Partial<T>;

type KeysOfUnion<T> = T extends T ? keyof T : never;
export type Exact<P, I extends P> = P extends Builtin ? P
  : P & { [K in keyof P]: Exact<P[K], I[K]> } & { [K in Exclude<keyof I, KeysOfUnion<P>>]: never };

function isSet(value: any): boolean {
  return value !== null && value !== undefined;
}

export interface MessageFns<T> {
  fromJSON(object: any): T;
  toJSON(message: T): unknown;
  create<I extends Exact<DeepPartial<T>, I>>(base?: I): T;
  fromPartial<I extends Exact<DeepPartial<T>, I>>(object: I): T;
}